Added modal logic

+++ nomnom/views.py
@@ -43,9 +43,12 @@ def delete_poll(poll_id, delete_key):
         flask.abort(404)
     if poll.delete_key != delete_key:
         flask.abort(403)
-    poll.key.delete()
-    flask.flash('Poll deleted successfully.', 'success')
-    return flask.redirect('/', code=302)  # Redirect back to home page
+    if flask.request.method == 'POST':
+        poll.key.delete()
+        flask.flash('Poll deleted successfully.', 'success')
+        return flask.redirect('/', code=302)  # Redirect back to home page
+    form = forms.ResponseForm()
+    return flask.render_template('poll.html', title=poll.title, poll=poll, responses=poll.get_responses(), form=form, delete=True)
 
 # Vote on a response to a poll
 @app.route('/poll/<string:poll_id>/vote/<string:vote_type>', methods=['POST'])
